app/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-autoslug==1.9.9
django-uuslug==2.0.0
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
